+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 
 - Add field type textarea
 - Update UI configuration for large attribute title
+- Fix #1 #2