+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Add field type textarea
 - Update UI configuration for large attribute title
 - Fix #1 #2
+- add static method getFirstValues in model AttributeType