+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Update UI configuration for large attribute title
 - Fix #1 #2
 - add static method getFirstValues in model AttributeType
+- Add field type boolean