+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - Fix #1 #2
 - add static method getFirstValues in model AttributeType
 - Add field type boolean
+- Add input args (attribute_type_id, attribute_type_slug) for loop attribute_availability_extend_attribute_type