+++ CHANGELOG.md
@@ -14,3 +14,7 @@
 - Add field type boolean
 - Add input args (attribute_type_id, attribute_type_slug) for loop attribute_availability_extend_attribute_type
 - Add input arg attribute_type_slug for loop attribute_extend_attribute_type
+
+# 1.1.1
+
+- fix returns values depending on the language